python/requirements.txt
Flask==2.2.3
gunicorn==20.1.0
mysql-connector-python==8.0.32
PyJWT==2.4.0
autopep8==2.0.2
pylint==2.16.4
serverless-wsgi==3.0.2
awslambdaric==2.0.4
Werkzeug==2.2.3
